Add files via upload
ATM Machine
This simple ATM System project is written in Python. The project file contains a python script (atm1.py). This is a simple
console based system which is very easy to use. Talking about the system, it contains Some functions which includes
Withdrawing,Depositing, and Show balance amount. It is easy to use, he/she just has to enter the amount whether in
deposit section or the withdraw section. The system calculates the total remaining balance of the respective account and
displays to the user.